Исправляет ошибки домашнего задания module5-task2
