+++ project/src/utils/utils.ts
@@ -1,0 +1,4 @@
+export function capitalize(str: string): string {
+  return `${str[0].toUpperCase()}${str.slice(1)}`;
+}
+
